requirements.txt
PyGithub
torch
librosa
